Extract macro to instantiate ring switch tests
